(PC-29257) refactor(subscription): group modals and button to prevent rerender
moving the modals state down the tree prevent flashes when a user
subscribes to a home thematic.
