+++ scripts/dead_code_snapshot.txt
@@ -1,2 +1,4 @@
 src/features/auth/helpers/contactSupport.ts:38 - satisfies
 src/features/auth/helpers/contactSupport.ts:38 - ContactSupport (used in module)
+src/features/home/fixtures/homepage.fixture.ts:206 - satisfies
+src/features/home/fixtures/homepage.fixture.ts:206 - Homepage (used in module)
