+++ src/features/subscription/helpers/useThematicSubscription.tsx
@@ -4,30 +4,31 @@ import { Platform } from 'react-native'
 import { UserProfileResponse } from 'api/gen'
 import { useUpdateProfileMutation } from 'features/profile/api/useUpdateProfileMutation'
 import { usePushPermission } from 'features/profile/pages/NotificationSettings/usePushPermission'
+import { useMapSubscriptionHomeIdsToThematic } from 'features/subscription/helpers/useMapSubscriptionHomeIdsToThematic'
 import { SubscriptionTheme } from 'features/subscription/types'
 import { analytics } from 'libs/analytics'
 import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 
 export type Props = {
   user?: UserProfileResponse
-  thematic: SubscriptionTheme | null
   homeId: string
-  onUpdateSubscriptionSuccess: () => Promise<void>
+  onUpdateSubscriptionSuccess: (thematic: SubscriptionTheme) => Promise<void>
 }
 
 export const useThematicSubscription = ({
   user,
-  thematic,
   homeId,
   onUpdateSubscriptionSuccess,
 }: Props): {
   isSubscribeButtonActive: boolean
   isAtLeastOneNotificationTypeActivated: boolean
+  thematic: SubscriptionTheme | null
   updateSubscription: () => void
   updateSettings: ({ allowEmails, allowPush }: { allowEmails: boolean; allowPush: boolean }) => void
 } => {
   const { pushPermission } = usePushPermission()
   const isPushPermissionGranted = pushPermission === 'granted'
+  const thematic = useMapSubscriptionHomeIdsToThematic(homeId)
 
   const { showErrorSnackBar } = useSnackBarContext()
 
@@ -54,9 +55,9 @@ export const useThematicSubscription = ({
   const { mutate: updateProfile, isLoading: isUpdatingProfile } = useUpdateProfileMutation(
     async () => {
       analytics.logNotificationToggle(!!state.allowEmails, !!state.allowPush)
-      if (!isSubscribeButtonActive) {
+      if (!isSubscribeButtonActive && thematic) {
         analytics.logSubscriptionUpdate({ from: 'thematicHome', type: 'in', entryId: homeId })
-        await onUpdateSubscriptionSuccess?.()
+        await onUpdateSubscriptionSuccess?.(thematic)
       } else {
         analytics.logSubscriptionUpdate({ from: 'thematicHome', type: 'out', entryId: homeId })
       }
@@ -74,6 +75,7 @@ export const useThematicSubscription = ({
     return {
       isSubscribeButtonActive: false,
       isAtLeastOneNotificationTypeActivated: false,
+      thematic,
       updateSubscription: () => 0,
       updateSettings: () => 0,
     }
@@ -114,6 +116,7 @@ export const useThematicSubscription = ({
   return {
     isSubscribeButtonActive: !!isSubscribeButtonActive,
     isAtLeastOneNotificationTypeActivated: !!isAtLeastOneNotificationTypeActivated,
+    thematic,
     updateSubscription,
     updateSettings,
   }
